requirements.txt

streamlit==1.32.2
streamlit-chat==0.1.1


openai==0.28.0

pdfplumber



python-dotenv 





